Added tests for 1,3 tasks

+++ app/src/main/java/org/tbank/controller/EventController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import org.tbank.dto.events.EventDTO;
 import org.tbank.service.EventService;
 import reactor.core.publisher.Mono;
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -20,7 +21,7 @@ public class EventController {
     public Mono<List<EventDTO>> getEvents(
             @RequestParam(value = "dateFrom", required = false) String dateFrom,
             @RequestParam(value = "dateTo", required = false) String dateTo,
-            @RequestParam(value = "budget") double budget,
+            @RequestParam(value = "budget") BigDecimal budget,
             @RequestParam(value = "currency") String currency) {
 
         return eventService.fetchEvents(dateFrom, dateTo, budget, currency);
